+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Changed "Reforge Item" effect to be able to reforge any valid item.
 - Changed "Increase Spawn Rate" effect to allow enemies to spawn in towns.
+- Changed "Spawn Honey Trap" effect to spawn a smaller amount of honey.
 - Fixed "Catch Critter" challenge not removing bug net from hand in some cases.
 
 ## [2.0.5] - 2023-09-20